# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Remont API
  version: 1.0.0
paths:
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_login_api_v1_auth_login_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: Read Users Me
      operationId: read_users_me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/machines/:
    get:
      tags:
      - machines
      summary: Read Machines
      operationId: read_machines_api_v1_machines__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 25
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MachineOut'
                title: Response Read Machines Api V1 Machines  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - machines
      summary: Create Machine
      operationId: create_machine_api_v1_machines__post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MachineCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/machines/{code}:
    get:
      tags:
      - machines
      summary: Read Machine
      operationId: read_machine_api_v1_machines__code__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/machines/import:
    post:
      tags:
      - machines
      summary: Import Machines
      operationId: import_machines_api_v1_machines_import_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: dry_run
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Dry Run
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_machines_api_v1_machines_import_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/machines/qr-sheet:
    post:
      tags:
      - machines
      summary: Generate Qr Sheet
      operationId: generate_qr_sheet_api_v1_machines_qr_sheet_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QrSheetRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/machines/{code}/qr:
    get:
      tags:
      - machines
      summary: Generate Qr Single
      operationId: generate_qr_single_api_v1_machines__code__qr_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/analytics/history:
    get:
      tags:
      - analytics
      summary: Get Analytics History
      operationId: get_analytics_history_api_v1_analytics_history_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: machine_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Machine Code
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date From
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date To
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsHistoryOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/analytics/dashboard:
    get:
      tags:
      - analytics
      summary: Get Analytics Dashboard
      operationId: get_analytics_dashboard_api_v1_analytics_dashboard_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: machine_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Machine Code
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date From
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date To
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AnalyticsHistoryOut:
      properties:
        repairs_count:
          type: integer
          title: Repairs Count
        new_parts_qty:
          type: number
          title: New Parts Qty
        new_parts_amount:
          anyOf:
          - type: number
          - type: 'null'
          title: New Parts Amount
        refurbished_parts_qty:
          type: number
          title: Refurbished Parts Qty
        top_parts:
          items:
            type: object
          type: array
          title: Top Parts
        refurbishment_share_pct:
          type: number
          title: Refurbishment Share Pct
        zaprafka_expenses:
          anyOf:
          - type: number
          - type: 'null'
          title: Zaprafka Expenses
      type: object
      required:
      - repairs_count
      - new_parts_qty
      - new_parts_amount
      - refurbished_parts_qty
      - top_parts
      - refurbishment_share_pct
      - zaprafka_expenses
      title: AnalyticsHistoryOut
    Body_import_machines_api_v1_machines_import_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_import_machines_api_v1_machines_import_post
    Body_login_api_v1_auth_login_post:
      properties:
        grant_type:
          anyOf:
          - type: string
            pattern: ^password$
          - type: 'null'
          title: Grant Type
        username:
          type: string
          title: Username
        password:
          type: string
          format: password
          title: Password
        scope:
          type: string
          title: Scope
          default: ''
        client_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
          - type: string
          - type: 'null'
          format: password
          title: Client Secret
      type: object
      required:
      - username
      - password
      title: Body_login_api_v1_auth_login_post
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MachineCreate:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        serial_no:
          anyOf:
          - type: string
          - type: 'null'
          title: Serial No
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        sap_co_order:
          anyOf:
          - type: string
          - type: 'null'
          title: Sap Co Order
        sap_cost_center:
          anyOf:
          - type: string
          - type: 'null'
          title: Sap Cost Center
        zaprafka_interval_months:
          type: integer
          title: Zaprafka Interval Months
          default: 60
      type: object
      required:
      - code
      - name
      title: MachineCreate
    MachineOut:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        serial_no:
          anyOf:
          - type: string
          - type: 'null'
          title: Serial No
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        sap_co_order:
          anyOf:
          - type: string
          - type: 'null'
          title: Sap Co Order
        sap_cost_center:
          anyOf:
          - type: string
          - type: 'null'
          title: Sap Cost Center
        zaprafka_interval_months:
          type: integer
          title: Zaprafka Interval Months
          default: 60
        id:
          type: string
          format: uuid
          title: Id
        status:
          type: string
          title: Status
        last_zaprafka_end:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Zaprafka End
      type: object
      required:
      - code
      - name
      - id
      - status
      title: MachineOut
...
    QrSheetRequest:
      properties:
        codes:
          items:
            type: string
          type: array
          title: Codes
      type: object
      required:
      - codes
      title: QrSheetRequest
...
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
        user:
          $ref: '#/components/schemas/UserOut'
      type: object
      required:
      - access_token
      - refresh_token
      - token_type
      - user
      title: Token
...
    UserOut:
      properties:
        username:
          type: string
          title: Username
        full_name:
          type: string
          title: Full Name
        role:
          type: string
          title: Role
        is_active:
          type: boolean
          title: Is Active
          default: true
        id:
          type: string
          format: uuid
          title: Id
      type: object
      required:
      - username
      - full_name
      - role
      - id
      title: UserOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: api/v1/auth/login